fix an incorrect IO error reading SMART status
The read SMART status command's return status was testing
for a success/failure value that included 8 bits that are
"N/A" according to the standard, and required that they be
zeros.  At least some drives do not fill them with zeros,
so correct this by masking off the undefined bits.

Signed-off-by: Phillip Susi <psusi@ubuntu.com>